Added DealController and DealService

+++ src/main/java/ua/shpp/service/ScheduleEventService.java
@@ -1,10 +1,10 @@
 package ua.shpp.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import ua.shpp.dto.ScheduleEventDto;
-import ua.shpp.exception.EventTypeNotFoundException;
-import ua.shpp.exception.OfferNotFoundException;
+import ua.shpp.exception.EventNotFoundException;
 import ua.shpp.mapper.ScheduleEventMapper;
 import ua.shpp.repository.ScheduleEventRepository;
 import ua.shpp.repository.ServiceRepository;
@@ -12,6 +12,7 @@ import ua.shpp.repository.ServiceRepository;
 import java.time.LocalDate;
 import java.util.List;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ScheduleEventService {
@@ -20,23 +21,39 @@ public class ScheduleEventService {
     private final ServiceRepository serviceRepository;
 
     public ScheduleEventDto create(ScheduleEventDto scheduleEventDto) {
-        var entity = scheduleEventMapper.toEntity(scheduleEventDto,serviceRepository);
+        log.debug("create() called with DTO: {}", scheduleEventDto);
+        var entity = scheduleEventMapper.toEntity(scheduleEventDto, serviceRepository);
         entity.setId(null);
         entity = scheduleEventRepository.save(entity);
-
+        log.info("Created schedule event (id={})", entity.getId());
+        log.debug("create() schedule event Entity: {}", entity);
         return scheduleEventMapper.toDto(entity);
     }
 
-    public ScheduleEventDto getById (Long id){
-        var entity = scheduleEventRepository.findById(id).orElseThrow(() -> new RuntimeException(String.format("Event id: %d, not found", id)));
+    public ScheduleEventDto getById(Long id) {
+        log.debug("get() called with id: {}", id);
+        var entity = scheduleEventRepository.findById(id)
+                .orElseThrow(() -> new EventNotFoundException(String.format("Event id: %d, not found", id)));
+        log.info("Fetching schedule event (id={})", id);
+        log.debug("Fetching schedule event entity: {}", entity);
         return scheduleEventMapper.toDto(entity);
     }
 
-    public List<ScheduleEventDto> getAllBetweenDates(LocalDate startDate, LocalDate endDate){
+    public List<ScheduleEventDto> getAllBetweenDates(LocalDate startDate, LocalDate endDate) {
+        log.debug("getAllBetweenDates() called with from {} to {} dates", startDate, endDate);
         var listEntity = scheduleEventRepository.findByEventDateBetween(startDate, endDate);
+        log.debug("dates between {} to {} have {} events",
+                startDate, endDate, listEntity.size());
         return listEntity.stream().map(scheduleEventMapper::toDto).toList();
     }
-    public void deleteById(Long id){
-        scheduleEventRepository.deleteById(id);
+
+    public void deleteById(Long id) {
+        log.debug("delete() called with id: {}", id);
+        if (!scheduleEventRepository.existsById(id)) {
+            log.debug("Event id: {}, not found", id);
+        } else {
+            scheduleEventRepository.deleteById(id);
+            log.info("Deleted event (id={})", id);
+        }
     }
 }
